requirements.txt
asgiref==3.9.1
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
deep-translator==1.11.4
Django==5.2.6
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-translate==3.21.1
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
idna==3.10
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
rsa==4.9.1
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
